requirements.txt
aiosqlite==0.22.1
alembic==1.17.2
anyio==4.12.1
asyncio==4.0.0
asyncpg==0.31.0
fastapi==0.128.0
graphql-core==3.2.7
httpx==0.28.1
idna==3.11
pydantic==2.12.5
pydantic-settings==2.12.0
pytest==9.0.2
pytest-asyncio==1.3.0
SQLAlchemy==2.0.45
starlette==0.50.0
strawberry-graphql==0.288.2
uvicorn==0.40.0
dotenv==0.9.9